finished porting important functionality wip:  ime on linux, better dialog, ensure MacOS compile, swaping in MMW macros...

+++ MikuMikuWorld/ScoreConverter.cpp
@@ -26,7 +26,7 @@ namespace MikuMikuWorld
 		for (int i = 2; i < 10; ++i)
 		{
 			if (fmodf(factor * length, 1) == 0)
-				return std::pair<int, int>(factor * length, pow(2, i));
+				return std::pair<int, int>((int)(factor * length), (int)pow(2, i));
 
 			factor *= 2;
 		}
